+++ include/diplib/private/pcg_uint128.hpp
@@ -170,7 +170,7 @@ inline bitcount_t trailingzeros(uint32_t v)
       31, 27, 13, 23, 21, 19, 16, 7, 26, 12, 18, 6, 11, 5, 10, 9
     };
 
-    return multiplyDeBruijnBitPos[((uint32_t)((v & -v) * 0x077CB531U)) >> 27];
+    return multiplyDeBruijnBitPos[((uint32_t)((v & -(int32_t)v) * 0x077CB531U)) >> 27];
 }
 
 inline bitcount_t flog2(uint64_t v)
